Updating unit tests.
