+++ test/unit/Sidebar.test.js
@@ -31,16 +31,6 @@ describe('Sidebar', () => {
       expect(sidebarWrapper.hasClass('nypl-column-one-quarter')).to.equal(true);
     });
 
-    it('should render a breadcrumb link', () => {
-      const breadcrumbLink = component.find('a').at(0);
-      expect(breadcrumbLink).to.have.length(1);
-      // The first part of the "text" is the hidden SVG title which is not hidden in the tests.
-      expect(breadcrumbLink.text())
-        .to.equal('NYPL Left Wedge SVG IconRecommendations');
-      expect(breadcrumbLink.prop('href'))
-        .to.equal('https://www.nypl.org/books-music-movies/recommendations');
-    });
-
     it('should render a <BookFilters /> component', () => {
       expect(component.find('BookFilters')).to.have.length(1);
     });
